requirements.txt
Django==1.4.1
MySQL-python==1.2.3
South==0.7.6
django-bootstrap-toolkit==2.5.6
PIL
markdown
facebook-sdk
